+++ database/migrations/2025_03_18_133820_create_detail_qc_po_table.php
@@ -13,14 +13,16 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::create('detail_qc_po', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('id_header_po')->references('po_id')->on('list_header_po')->onDelete('cascade');
-            $table->string('nip_user')->references('nip')->on('users')->onDelete('cascade');
-            $table->string('keterangan');
-            $table->string('status');
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('detail_qc_po')) {
+            Schema::create('detail_qc_po', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('id_header_po')->references('po_id')->on('list_header_po')->onDelete('cascade');
+                $table->string('nip_user')->references('nip')->on('users')->onDelete('cascade');
+                $table->string('keterangan');
+                $table->string('status');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
