+++ database/migrations/2025_03_18_141624_update_column_detail_qc_po_table.php
@@ -14,9 +14,13 @@ return new class extends Migration
     public function up()
     {
         Schema::table('detail_qc_po', function (Blueprint $table) {
-            $table->dropColumn('nip');
+            if (Schema::hasColumn('detail_qc_po', 'nip')) {
+                $table->dropColumn('nip');
+            }
             $table->string('nip_user')->references('nip')->on('users')->onDelete('cascade');
-            $table->dropColumn('id_header_po');
+            if (Schema::hasColumn('detail_qc_po', 'id_header_po')) {
+                $table->dropColumn('id_header_po');
+            }
             $table->foreignId('id_header_po')->references('po_id')->on('list_header_po')->onDelete('cascade');
         });
     }
