+++ database/migrations/2025_03_19_100932_update_column2_detail_qc_po_table.php
@@ -15,7 +15,9 @@ return new class extends Migration
     {
         Schema::table('detail_qc_po', function (Blueprint $table) {
             //$table->dropForeign(['id_header_po']);
-            $table->dropColumn('id_header_po');
+            if (Schema::hasColumn('detail_qc_po', 'id_header_po')) {
+                $table->dropColumn('id_header_po');
+            }
         });
     }
